Merge adjacent text nodes in InlineParser
With this change, a text node should never have another text node as an
adjacent sibling. In other words, text nodes span as much literal text
as possible.

This is useful for post-processing tasks such as autolinking. It can be
done externally using a visitor, but doing it as we go seems cleaner.

Benchmarking didn't show a change in performance with this change.