gfm-tables: Support single-column tables (#7)
Contrary to what GitHub does, the following is not turned into a table:

    |Header
    ---

Because we require the separator line to have at least one pipe in it
(either at the start or the end or between columns). Otherwise it
conflicts with a setext heading. If this turns out to be a problem, we
can always revisit this.