Make AttributeProvider work for Image nodes (#31)
Instead of disabling tags in the writer and using the same visitor for
rendering the alt text, just use a separate visitor.

This allows us to get rid of some complicated logic too. The only
downside is that we now have to build up the whole alt text in memory.