Make HTML rendering for nodes extensible (#35)
We had `CustomHtmlRenderer` for custom nodes before. That was ok for
that, but didn't allow clients to override rendering for core nodes.

This introduces a new concept `NodeRenderer` (including factories, who
doesn't like those?) that is also used for the core nodes. That allows
overriding the rendering and allows to define the rendering for custom
nodes at the same time.

NodeRenderers have access to rendering configuration and functionality,
such as extending attributes. That fixes the problem that the renderer
for table nodes wasn't extensible via `AttributeProvider` (#31).