Fix StringIndexOutOfBoundsException on line after tab (#52)
This could occur in the following situation (\t is a tab):

    - foo

    \tbar
    # baz

The tab on the bar line is a partially consumed tab, so columnIsInTab
would be true. On the next line, we wouldn't reset columnIsInTab.

By the time that addLine is called, we would try to get a substring
starting from "after the tab", resulting in the
StringIndexOutOfBoundsException.