Simplify merging of adjacent text nodes
Before, we would merge them while parsing. It was very complicated
because we had to be careful to not add other nodes if we still had
outstanding text nodes.

Merging them at the end makes things simpler. It also means we can
allocate a StringBuilder with the final size instead of having to expand
it.

We could make it even simpler by using a visitor at the very end of
parsing, but I like the current approach because it doesn't require
descending into child nodes.

What motivated this change is work on source positions for inline nodes.
This change should make it simpler to implement.