Simplify and speed up brackets processing (links/images)
The key insight for the simplification is that we don't need to keep the
brackets in the same stack as the other delimiters. All we need to know
is which delimiter came before the bracket, so that we can process the
emphasis in the image/link text.

A nice side effect of this is that we don't need to go through the
delimiters when looking for the last opening bracket, we already have
it. This means we can drop the `bracketDelimiterBottom` optimization.

And by remembering the `bracketAfter` on the last bracket, we can
immediately discard matched brackets instead of marking them as
`matched` and keeping them in the stack.

All this together allows us to properly fix the `nestedBrackets`
pathological case, yay :).