Move new emphasis rule into DelimiterProcessor (extension api break)
The "multiple of 3" rule was breaking the strikethrough extension logic.

By passing more information into getDelimiterUse and allowing the
implementation to return 0, we can move the "multiple of 3" rule into
the delimiter processor.

For the strikethrough delimiter processor, we can have different rules
and simplify the implementation. It's now also closer to GFM.

Also clean up the other method names a bit while we are breaking the
API:

* getOpeningDelimiterChar -> getOpeningCharacter
* getClosingDelimiterChar -> getClosingCharacter
* getMinDelimiterCount -> getMinLength