Add new tests to check for unicode support
An original assumption was that the id should be ascii. This assumption
has been removed and support should be added for international
charcaters and diacritics. Rather than attempt to support all combining
characters, tests are written to ensure that there is support for the
diacritical marks. Namely those in the Combining Diactirical Marks
unicode, specifically (0x0300-0x036F)