Support multiple DelimiterProcessors with the same delimiter char
Previously it was not possible to use the same delimiter char for
different DelimiterProcessors even if they require a different number of
delimiters (for example, using `~` for subscript together with `~~` for
strikethrough). This change adds support for multiple
DelimiterProcessors with the same delimiter char as long as all
DelimiterProcessors for that char have the same opening and closing
char and different minimum lengths. When multiple processors are defined
for a char, the one with the largest acceptable minimum length is
chosen in each case.