Adapt link reference definition parsing to spec changes
The spec was updated to clarify that they can also be part of setext
headings. The way we make these work in our parser is to be able to
"replace" the active paragraph block. In that case, we still need to
collect link reference definitions from it.

In order to implement this, I had to separate it from InlineParser. The
new way is cleaner, and allows us to add a Node to the document as well
(for #98).

I think there's still room for improvement: We could parse the
definition as we go in ParagraphParser and collect them earlier. That
might eliminate the current "double parsing" that we do in some cases.