Start extracting inline parsers
The end goal is to extract all current inline parsers, and then allow
customizing inline parsing via the same infrastructure (similar to the
current block parsing).

A stretch goal is to move away from the parser input having to be a
single contiguous String. This could speed up parsing and allow keeping
track of source positions (line and column indexes) for inline nodes.