Restart work for adding source positions
The previous attempt required every block parser to implement source
positions, which is not ideal. This one is for trying to see if we can
get it done inside DocumentParser.

Note that SourceSpan initially used 1-based indexes and
beginIndex/endIndex, but it was changed to 0-based and a length because:

0-based fits better with the existing API and other Java APIs. And
instead of having an endIndex, having a length would allow us to easily
extend the SourceSpan to include a character offset in addition to
line/column offset.