Set block source spans based on the index
This means that e.g. this:

```
 paragraph starting with a space
```

Will include the starting space in the source positions. I was initially
thinking it should not be included, but that makes a lot of things
harder, for not much gain IMO.

E.g. for a code block with 5 spaces:

```
     code
```

Where should the source position start, at index 0, 4, or 5? This
implements 0. 4 is kind of hard to implement due to tabs being treated
as 4-space tabstops.