Add `requireTwoTildes` for `StrikethroughExtension` (fixes #271)
With the previous version we adjusted the extension to also accept the
single tilde syntax. But if you use another extension that uses the
single tilde syntax, you will get a conflict, see issue.

To avoid that, `StrikethroughExtension` can now be configured to require
two tildes like before, see Javadoc.