add column width to TableCell nodes
The width is determined by the number of dash and colon characters in the
separator line of a table, and passed on to the consumer as an integer.
This permits consumers of the AST to render tables that are more
proportional the width in the source markdown.

No changes have been made to the HTML render to make use of the width
information, but the test for this feature demonstrates a possible usage.

Should a cell be be created that is further to the right than the parsed
separator columns are (i.e., when alignment would be null), then the width
will be reported as 0.