Fix link reference definition parsing with invalid title
An input like this:

```
[foo]: /url
"title" bad
```

Means the second line is just a paragraph because only spaces/tabs are allowed
after a title. The parser used to set the title to "title" in this case and
assign the source span of the second line to the definition, which is wrong.

Fixes #315.