Support nested footnotes
This turned out to be tricky, and GitHub gets some of it wrong.
If anyone ever wants us to be bug-compatible, it should be relatively
straightforward to emulate GitHub by just running the initial reference
search over everything (including definitions) and then not bothering
with finding more at the end.