Generalize link markers and use for inline footnotes
The resulting implementation for the footnotes extension is much nicer.
It also cleans up LinkInfo and makes images less of a special case.

Additionally, this allow inline parsing of markers that are not
part of links - could have done this without this change but noticed
it here and decided to fix it.